#1 Refactor and remove unnecessary files
