+++ app/src/main/java/co/thiennguyen/github_profile/di/modules/AppModule.kt
@@ -1,7 +1,7 @@
 package co.thiennguyen.github_profile.di.modules
 
-import co.thiennguyen.github_profile.DispatchersProvider
-import co.thiennguyen.github_profile.DispatchersProviderImpl
+import co.thiennguyen.github_profile.util.DispatchersProvider
+import co.thiennguyen.github_profile.util.DispatchersProviderImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -10,6 +10,7 @@ import dagger.hilt.components.SingletonComponent
 @Module
 @InstallIn(SingletonComponent::class)
 class AppModule {
+
     @Provides
     fun provideDispatchersProvider(): DispatchersProvider {
         return DispatchersProviderImpl()
