+++ app/src/main/java/co/thiennguyen/github_profile/ui/base/BaseDestination.kt
@@ -6,8 +6,6 @@ abstract class BaseDestination(val route: String = "") {
 
     open val arguments: List<NamedNavArgument> = emptyList()
 
-    open val deepLinks: List<String> = emptyList()
-
     open var destination: String = route
 
     data class Up(val results: HashMap<String, Any> = hashMapOf()) : BaseDestination() {
