+++ app/src/main/java/co/thiennguyen/github_profile/ui/base/BaseScreen.kt
@@ -8,6 +8,5 @@ fun BaseScreen(
     content: @Composable () -> Unit,
 ) {
     // TODO Base logic for all screens here
-
     content()
 }
