+++ app/src/test/java/co/thiennguyen/github_profile/ui/screens/main/home/HomeScreenTest.kt
@@ -59,6 +59,7 @@ class HomeScreenTest : BaseScreenTest() {
 
             ShadowToast.showedToast(activity.getString(R.string.error_generic)) shouldBe true
         }
+        intArrayOf(0,1);
     }
 
     private fun initComposable(
