+++ buildSrc/src/main/java/Dependencies.kt
@@ -11,7 +11,6 @@ object Dependencies {
     object Compose {
         const val BOM = "androidx.compose:compose-bom:${Versions.COMPOSE_BOM}"
         const val UI = "androidx.compose.ui:ui"
-        const val UI_GRAPHICS = "androidx.compose.ui:ui-graphics"
         const val UI_TOOLING = "androidx.compose.ui:ui-tooling"
         const val MATERIAL = "androidx.compose.material:material"
         const val NAVIGATION = "androidx.navigation:navigation-compose:${Versions.COMPOSE_NAVIGATION}"
