+++ data/src/main/java/co/thiennguyen/github_profile/data/remote/providers/MoshiBuilderProvider.kt
@@ -10,7 +10,6 @@ object MoshiBuilderProvider {
     val moshiBuilder: Moshi.Builder
         get() = Moshi.Builder()
             // Parse the DateTime in this format: [yyyy-MM-ddThh:mm:ss.ssZ]
-            // e.g: [2019-10-12T07:20:50.52Z]
             .add(Date::class.java, Rfc3339DateJsonAdapter())
             .add(KotlinJsonAdapterFactory())
 }
