+++ env.properties
@@ -1,0 +1,1 @@
+BASE_API_URL="https://BASE_API_URL"
